Split attention CLI commands
