+++ attention.ts
@@ -1,179 +1,4 @@
 #!/usr/bin/env bun
-import { existsSync } from "node:fs";
-import { cp, mkdir, rm } from "node:fs/promises";
-import path from "node:path";
-import { attentionDataDir, attentionDbPath, attentionHome } from "./server/paths";
-import { LocalSqliteStore } from "./server/sqlite";
+import { runAttentionCli } from "./server/cli";
 
-const rawArgs = process.argv.slice(2);
-if (rawArgs[0] === "--") rawArgs.shift();
-const [command = "help", subcommand, ...rest] = rawArgs;
-const apiPort = Number(process.env.ATTENTION_API_PORT ?? 4332);
-type DoctorCheck = { name: string; ok: boolean; detail: string };
-
-function print(value: unknown): void {
-  process.stdout.write(`${typeof value === "string" ? value : JSON.stringify(value, null, 2)}\n`);
-}
-
-async function initRuntime(): Promise<LocalSqliteStore> {
-  await mkdir(attentionHome(), { recursive: true });
-  await mkdir(attentionDataDir(), { recursive: true });
-  const sqlite = new LocalSqliteStore();
-  await sqlite.init();
-  return sqlite;
-}
-
-async function status(): Promise<void> {
-  const sqlite = await initRuntime();
-  print({
-    home: attentionHome(),
-    dataDir: attentionDataDir(),
-    dbPath: attentionDbPath(),
-    apiUrl: `http://127.0.0.1:${apiPort}`,
-    uiUrl: `http://127.0.0.1:${apiPort}`,
-    mcpUrl: `http://127.0.0.1:${apiPort}/mcp`,
-    sqlite: sqlite.status(),
-  });
-  sqlite.close();
-}
-
-async function doctor(): Promise<void> {
-  const sqlite = await initRuntime();
-  const status = sqlite.status();
-  const checks = [
-    { name: "home", ok: existsSync(attentionHome()), detail: attentionHome() },
-    { name: "data directory", ok: existsSync(attentionDataDir()), detail: attentionDataDir() },
-    { name: "sqlite database", ok: existsSync(attentionDbPath()) && status.schemaVersion >= 1, detail: `${attentionDbPath()} schema=${status.schemaVersion}` },
-    await checkApiStatus(),
-  ];
-  print({ ok: checks.every((check) => check.ok), checks });
-  sqlite.close();
-}
-
-async function checkApiStatus(): Promise<DoctorCheck> {
-  const url = `http://127.0.0.1:${apiPort}/api/status`;
-  const controller = new AbortController();
-  const timeout = setTimeout(() => controller.abort(), 1000);
-  try {
-    const response = await fetch(url, { signal: controller.signal });
-    if (!response.ok) return { name: "local api", ok: false, detail: `${url} returned HTTP ${response.status}` };
-    const status = await response.json() as { ok?: boolean; mcpUrl?: string; sqlite?: { schemaVersion?: number } };
-    const mcpUrl = `http://127.0.0.1:${apiPort}/mcp`;
-    const ok = status.ok === true && status.mcpUrl === mcpUrl && Number(status.sqlite?.schemaVersion ?? 0) >= 1;
-    return { name: "local api", ok, detail: ok ? `${url} reachable; MCP ${mcpUrl}` : `${url} returned an unexpected status payload` };
-  } catch (error) {
-    const reason = error instanceof Error && error.name === "AbortError" ? "timed out" : "not reachable";
-    return { name: "local api", ok: false, detail: `${url} ${reason}. Run attention start, then rerun doctor.` };
-  } finally {
-    clearTimeout(timeout);
-  }
-}
-
-function setupCodex(): void {
-  const mcpUrl = `http://127.0.0.1:${apiPort}/mcp`;
-  print(`Add this MCP server to Codex Desktop:
-
-Name: attention
-URL: ${mcpUrl}
-
-Then start one fresh Codex thread per feed and use this prompt:
-
-Connect this Codex Desktop thread to local Attention.
-
-MCP server: ${mcpUrl}
-Feed: inbox
-
-Inspect the feed, bind this thread as the feed home thread with bind_feed_thread, and create or update one heartbeat automation on this same thread. On each wakeup, inspect the feed, list queued work first, claim before using local connectors for queued instructions, execute and complete/fail/block/retry/cancel each claim through Attention MCP, verify approved external actions immediately before mutation, and refresh configured sources only when no queued work is being handled.
-`);
-}
-
-async function backupExport(targetPath: string): Promise<void> {
-  const sqlite = await initRuntime();
-  sqlite.close();
-  await mkdir(path.dirname(targetPath), { recursive: true });
-  if (existsSync(targetPath)) await rm(targetPath, { recursive: true, force: true });
-  await mkdir(targetPath, { recursive: true });
-  await cp(attentionDataDir(), path.join(targetPath, "data"), { recursive: true });
-  if (existsSync(attentionDbPath())) await cp(attentionDbPath(), path.join(targetPath, "attention.db"));
-  await Bun.write(path.join(targetPath, "manifest.json"), JSON.stringify({
-    name: "attention-backup",
-    format: 2,
-    exportedAt: new Date().toISOString(),
-    dataDir: attentionDataDir(),
-    dbPath: attentionDbPath(),
-  }, null, 2));
-  print({ ok: true, exported: { dataDir: attentionDataDir(), dbPath: attentionDbPath() }, to: targetPath });
-}
-
-async function backupImport(sourcePath: string): Promise<void> {
-  if (!existsSync(sourcePath)) throw new Error(`Backup path does not exist: ${sourcePath}`);
-  const bundledData = path.join(sourcePath, "data");
-  const bundledDb = path.join(sourcePath, "attention.db");
-  const sourceData = existsSync(bundledData) ? bundledData : sourcePath;
-  await mkdir(attentionHome(), { recursive: true });
-  if (existsSync(attentionDataDir())) await rm(attentionDataDir(), { recursive: true, force: true });
-  await cp(sourceData, attentionDataDir(), { recursive: true });
-  await removeSqliteFiles();
-  if (existsSync(bundledDb)) await cp(bundledDb, attentionDbPath());
-  print({ ok: true, imported: sourcePath, to: { dataDir: attentionDataDir(), dbPath: attentionDbPath(), sqlite: existsSync(bundledDb) ? "restored" : "will_rehydrate_from_file_mirrors" } });
-}
-
-async function removeSqliteFiles(): Promise<void> {
-  await Promise.all([
-    rm(attentionDbPath(), { force: true }),
-    rm(`${attentionDbPath()}-shm`, { force: true }),
-    rm(`${attentionDbPath()}-wal`, { force: true }),
-  ]);
-}
-
-async function runLegacyCli(args: string[]): Promise<void> {
-  process.argv = [process.argv[0] ?? "bun", "cli.ts", ...args];
-  await import("./cli");
-}
-
-switch (command) {
-  case "start":
-    await initRuntime();
-    process.env.ATTENTION_CLIENT_DIR ??= path.join(process.cwd(), "dist");
-    print(`attention starting
-UI:  http://127.0.0.1:${apiPort}
-API: http://127.0.0.1:${apiPort}
-MCP: http://127.0.0.1:${apiPort}/mcp
-Data: ${attentionDataDir()}
-`);
-    await import("./server");
-    break;
-  case "status":
-    await status();
-    break;
-  case "doctor":
-    await doctor();
-    break;
-  case "setup":
-    if (subcommand !== "codex") throw new Error("Expected: attention setup codex");
-    setupCodex();
-    break;
-  case "backup":
-    if (subcommand === "export") await backupExport(rest[0] ?? path.join(attentionHome(), "exports", `attention-${Date.now()}`));
-    else if (subcommand === "import") await backupImport(rest[0] ?? "");
-    else throw new Error("Expected: attention backup export [path] or attention backup import <path>");
-    break;
-  case "help":
-    print({
-      commands: [
-        "attention start",
-        "attention status",
-        "attention doctor",
-        "attention setup codex",
-        "attention backup export [path]",
-        "attention backup import <path>",
-        "attention cli <existing-low-level-command> [...args]",
-      ],
-    });
-    break;
-  case "cli":
-    await runLegacyCli([subcommand, ...rest].filter((value): value is string => Boolean(value)));
-    break;
-  default:
-    await runLegacyCli([command, subcommand, ...rest].filter((value): value is string => Boolean(value)));
-}
+await runAttentionCli(process.argv.slice(2));
